requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cors==1.0.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2     
djangorestframework-simplejwt==5.3.1
django-import-export==2.5.0     # Added new package
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.0
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupytext==1.16.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.2                  # Updated version
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
tldextract==5.1.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
xgboost==2.1.1
zope.event==5.0
zope.interface==7.0.3
whitenoise==6.2.0
